tests/wpt/web-platform-tests/offscreen-canvas/drawing-rectangles-to-the-canvas/2d.strokeRect.negative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.strokeRect.negative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.strokeRect.negative</h1>
<p class="desc">strokeRect of negative sizes works</p>


<script>
var t = async_test("strokeRect of negative sizes works");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.strokeStyle = '#0f0';
ctx.lineWidth = 25;
ctx.strokeRect(12, 12, 26, 1);
ctx.strokeRect(88, 12, -26, 1);
ctx.strokeRect(12, 38, 26, -1);
ctx.strokeRect(88, 38, -26, -1);
_assertPixel(offscreenCanvas, 25,12, 0,255,0,255, "25,12", "0,255,0,255");
_assertPixel(offscreenCanvas, 75,12, 0,255,0,255, "75,12", "0,255,0,255");
_assertPixel(offscreenCanvas, 25,37, 0,255,0,255, "25,37", "0,255,0,255");
_assertPixel(offscreenCanvas, 75,37, 0,255,0,255, "75,37", "0,255,0,255");
t.done();

});
</script>
